Refactor DatabaseMemo model
